firewall_package/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
perspective-api-client==0.10.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
